chore: ViewModel 디렉토리 위치 오류 > 조정
