--- POME/Domain/ViewModel/Review/ReviewViewModel.swift
@@ -1,33 +1,0 @@
-//
-//  ReviewViewModel.swift
-//  POME
-//
-//  Created by 박지윤 on 2023/01/17.
-//
-
-import Foundation
-
-class ReviewViewModel{
-    
-//    private let reviewUseCase: GetReviewUseCase
-    
-    struct Input{
-        
-    }
-    
-    struct Output{
-        
-    }
-    
-    /*
-    init(reviewUseCase: GetReviewUseCase){
-
-    }
-     */
-    
-    func transform(input: Input) -> Output{
-        return Output()
-    }
-    
-    
-}
