+++ POME/Presentation/ViewModel/Record/Register/GenerateRecordViewModel.swift
